[Milan] Simplified the logic
